Create Toppers.cpp
Problem statement:

There is a class of N students and the task is to find the top K marks-scorers. Write a program that will print the index of the toppers of the class which will be same as the index of the student in the input array (use 0-based indexing). First print the index of the students having highest marks then the students with second highest and so on. If there are more than one students having same marks then print their indices in ascending order.
K=2
    N=10
    Marks are:
    97 84 82 89 84 97 95 95 84 86
    Output:
    0 5 6 7, so there are four toppers for K=2